fix field type for turbidity which was causing samples to be rejected

+++ src/Model/Table/WaterQualitySamplesTable.php
@@ -112,7 +112,7 @@
                 ->allowEmpty('DO_Exception');
 
             $validator
-                ->integer('Turbidity')
+                ->decimal('Turbidity')
                 ->allowEmpty('Turbidity');
 
             $validator
